+++ kmserver-core/src/dao/sqls/favorites.ts
@@ -1,12 +1,13 @@
 export const selectFavorites = `
-SELECT uf.fk_kid AS kid
+SELECT uf.fk_kid AS kid,
+	   uf.favorited_at AS favorited_at
 FROM  users_favorites uf
 WHERE uf.fk_login = $1
 `;
 
 export const insertFavorite = `
-INSERT INTO users_favorites (fk_login, fk_kid)
-VALUES($1, $2)
+INSERT INTO users_favorites (fk_login, fk_kid, favorited_at)
+VALUES($1, $2, $3)
 ON CONFLICT (fk_login, fk_kid) DO NOTHING;
 `;