+++ CHANGELOG.md
@@ -14,7 +14,7 @@ It's about time.
 - Maintainers: Maintainers can no remove suggestions
 - Import: You can now forbid creation of some tag types when importing songs via the import page. See sample config file for info.
 - Import: Number of inboxes submitted can be limited by user. This is a placeholder before we deploy the "My Submissions" feature subset. Check sample config file for help. (`Import.LoginNeeded` and `Import.MaxPerUser`)
-
+- Favorites: Favorited date is now added to the database from now on.
 
 ## Fixes